Make average rating a floating point value
* So 4.1 and 4.9 are not treated the same
* Currently has just 0.1 precision, so 4.95 and 4.9 are considered the same

+++ food-truck-api/src/main/java/food/truck/api/endpoint/TruckEndpoint.java
@@ -123,7 +123,7 @@ public class TruckEndpoint {
         @Nullable
         String description;
         @Nullable
-        Long priceRating;
+        Double priceRating;
         @Nullable
         String foodCategory;
         // TODO What about menu/schedule?
